Update README.md`
In my previous commit I assumed github.com using a browser to edit README.md wouldn't sign it. I was wrong, it can but not using my key, which it can't do as it doesn't have my private key, instead it signs with it's key.

This is "valid" because only :"I" can edit files based on keys I've previously installed in my account.